requirements.txt
Mastodon.py == 1.8.1
python-dotenv
